--- a/convert.py
+++ b/convert.py
@@ -275,7 +275,7 @@ class Params:
             n_ff = model["layers.0.feed_forward.experts.0.w1.weight"].shape[0]
             n_ctx = 32768
         else:
-            n_ff = model["layers.0.feed_forward.w1.weight"].shape[0],
+            n_ff = model["layers.0.feed_forward.w1.weight"].shape[0]
 
         return Params(
             n_vocab          = model["tok_embeddings.weight"].shape[0],