… ggml-metal.metal
… ggml.c
… ggml.h
… gguf-py/gguf/constants.py
… gguf-py/gguf/gguf_writer.py
… gguf-py/gguf/tensor_mapping.py
… llama.cpp
… tests/test-backend-ops.cpp