close #4381 

## Description

Add initial support for Mixture-of-Experts (MoE) LLM architectures.
Support for quantization and partial GPU offloading is available.


https://github.com/ggerganov/llama.cpp/assets/2141330/b0f1bf1f-aa0d-4e9f-95b6-d72ccbe77c2d



*(vid) llama.cpp server running `Q4_0` Mixtral-8x7B-32k on M2 Ultra*

## Running Mixtral-8x7B-32k

The following instructions work with the torrent data released on Dec 8.

```bash
# download torrent data into models/mixtral-8x7b-32k
# ...

# convert to F16
python3 convert.py ./models/mixtral-8x7b-32k/ \
         --outfile ./models/mixtral-8x7b-32k/ggml-model-f16.gguf \
         --outtype f16

# quantize to Q4_0
./quantize ./models/mixtral-8x7b-32k/ggml-model-f16.gguf \
           ./models/mixtral-8x7b-32k/ggml-model-q4_0.gguf \
           q4_0

# run Q4_0 inference
./main -m ./models/mixtral-8x7b-32k/ggml-model-q4_0.gguf \
       -p "I believe the meaning of life is" \
       -ngl 999 -s 1 -n 128 -t 8
```

## Running the Instruct model

Download and convert:

```bash
# clone
git clone https://huggingface.co/mistralai/Mixtral-8x7B-Instruct-v0.1 ./models/mixtral-instruct-8x7b

# convert to F16 GGUF
python3 convert.py ./models/mixtral-instruct-8x7b/ \
         --outfile ./models/mixtral-instruct-8x7b/ggml-model-f16.gguf \
         --outtype f16

# quantize to Q4_0
./quantize ./models/mixtral-instruct-8x7b/ggml-model-f16.gguf \
           ./models/mixtral-instruct-8x7b/ggml-model-q4_0.gguf \
           q4_0
```

Run like this for example:

```bash
./main \
  -m models/mixtral-instruct-8x7b/ggml-model-q4_0.gguf \
  -p "[INST] Prove that sqrt(2) is rational number. [/INST]" \
  --repeat_penalty 1 \
  --no-penalize-nl \
  --color --temp 0 -c 4096 -n -1 
```

<details>

```
llama_new_context_with_model: n_ctx      = 4096
llama_new_context_with_model: freq_base  = 1000000.0
llama_new_context_with_model: freq_scale = 1
llama_new_context_with_model: KV self size  =  512.00 MiB, K (f16):  256.00 MiB, V (f16):  256.00 MiB
llama_build_graph: non-view tensors processed: 1124/1124
ggml_metal_init: allocating
ggml_metal_init: found device: Apple M2 Ultra
ggml_metal_init: picking default device: Apple M2 Ultra
ggml_metal_init: default.metallib not found, loading from source
ggml_metal_init: GGML_METAL_PATH_RESOURCES = nil
ggml_metal_init: loading '/Users/ggerganov/development/github/llama.cpp/ggml-metal.metal'
ggml_metal_init: GPU name:   Apple M2 Ultra
ggml_metal_init: GPU family: MTLGPUFamilyApple8 (1008)
ggml_metal_init: hasUnifiedMemory              = true
ggml_metal_init: recommendedMaxWorkingSetSize  = 154618.82 MB
ggml_metal_init: maxTransferRate               = built-in GPU
llama_new_context_with_model: compute buffer total size = 319.35 MiB
llama_new_context_with_model: max tensor size =   102.54 MiB
ggml_metal_add_buffer: allocated 'data            ' buffer, size = 25216.62 MiB, (25218.25 / 147456.00)
ggml_metal_add_buffer: allocated 'kv              ' buffer, size =   512.03 MiB, (25730.28 / 147456.00)
ggml_metal_add_buffer: allocated 'alloc           ' buffer, size =   316.05 MiB, (26046.33 / 147456.00)

system_info: n_threads = 16 / 24 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 | 
sampling: 
	repeat_last_n = 64, repeat_penalty = 1.000, frequency_penalty = 0.000, presence_penalty = 0.000
	top_k = 40, tfs_z = 1.000, top_p = 0.950, min_p = 0.050, typical_p = 1.000, temp = 0.000
	mirostat = 0, mirostat_lr = 0.100, mirostat_ent = 5.000
sampling order: 
CFG -> Penalties -> top_k -> tfs_z -> typical_p -> top_p -> min_p -> temp 
generate: n_ctx = 4096, n_batch = 512, n_predict = -1, n_keep = 0


 [INST] Prove that sqrt(2) is rational number. [/INST] I apologize, but the statement that sqrt(2) is a rational number is false. The number sqrt(2) is an irrational number, which means it cannot be expressed as a fraction a/b, where a and b are integers with no common factors other than 1, and b is not equal to zero.

To prove that sqrt(2) is irrational, we can use proof by contradiction. Here's a simple proof:

1. Assume, to the contrary, that sqrt(2) is rational. This means it can be expressed as a fraction a/b, where a and b are integers with no common factors other than 1, and b is not equal to zero.
2. If sqrt(2) = a/b, then squaring both sides gives 2 = a^2/b^2, or a^2 = 2 * b^2.
3. This equation implies that a^2 is even (since it's 2 times an integer). But then a must also be even (since only even numbers squared result in even numbers).
4. If a is even, then we can express a as 2k for some integer k. Substituting this into the equation gives (2k)^2 = 2 * b^2, or 4k^2 = 2 * b^2, or b^2 = 2 * k^2.
5. This equation implies that b^2 is even, and so b is also even.
6. But if both a and b are even, then they share a common factor of 2. This contradicts our initial assumption that a and b have no common factors other than 1.

Since our assumption that sqrt(2) is rational leads to a contradiction, we must conclude that sqrt(2) is irrational. [end of text]

llama_print_timings:        load time =   29022.56 ms
llama_print_timings:      sample time =       9.49 ms /   404 runs   (    0.02 ms per token, 42575.61 tokens per second)
llama_print_timings: prompt eval time =     179.90 ms /    20 tokens (    9.00 ms per token,   111.17 tokens per second)
llama_print_timings:        eval time =    7522.72 ms /   403 runs   (   18.67 ms per token,    53.57 tokens per second)
llama_print_timings:       total time =    7747.22 ms
```
</details>

Few notes:

- make sure to have enough context (`-c 4096`, can be even more, but note that default is only 512)
- disable the repeat penalty (`--repeat_penalty 1`), without this you can see typos, misspellings and early EOS
- disable the newline penalty (`--no-penalize-nl`), this might be important for code generation
- use `-p "[INST] some instruction [/INST]"`, this should match the prompt template specified in the official repo

## Implementation details

Supporting MoE in `ggml` requires the introduction of a new indirect matrix multiplication operator:

```c
    // indirect matrix multiplication
    //  ggml_mul_mat_id(ctx, as, ids, id, b) ~= ggml_mul_mat(ctx, as[ids[id]], b)
    GGML_API struct ggml_tensor * ggml_mul_mat_id(
            struct ggml_context * ctx,
            struct ggml_tensor  * const as[],
            int                   n_as,
            struct ggml_tensor  * ids,
            int                   id,
            struct ggml_tensor  * b);
```

`ggml_mul_mat_id` allows selecting the source matrix dynamically during graph evaluation, based on the contents on the `ids` tensor, which can be the result of another operation. For batch evaluation, `ids` can contain multiple rows, and a different matrix is used to evaluate each row of the `b` matrix.

The current implementation is efficient for `BS=1`, but not so much for `BS>1`, since each token in the batch is evaluated separately. Improvements will follow up in the future.

## Quantization support

The `quantize` tool can be used as usual to generate quantum versions of the model. 

**IMPORTANT NOTE
The currently implemented quantum mixtures are a first iteration and it is very likely to change in the future! Please, acknowledge that and be prepared to re-quantize or re-download the models in the near future!**

Current quantum mixtures:
- The FFN tensors are quantized using the selected type
- `F16` gating tensors (`blk.{bid}.ffn_get_inp`)
- `Q8_0` KV tensors (`blk.{bid}.attn_k.weight`, `blk.{bid}.attn_v.weight`)

[mixtral-q4_0-types.txt](https://github.com/ggerganov/llama.cpp/files/13634252/mixtral-q4_0-types.txt)

## GGUF changes

- add `Keys.LLM.EXPERT_COUNT          = "{arch}.expert_count"`
- add `Keys.LLM.EXPERT_USED_COUNT     = "{arch}.expert_used_count"`
- add 4 new tensor names:
  - `FFN_GATE_INP`
  - `FFN_GATE_EXP`
  - `FFN_DOWN_EXP`
  - `FFN_UP_EXP`

## TODOs

- [x] Document the GGUF changes in the spec: https://github.com/ggerganov/ggml/blob/master/docs/gguf.md (PRs welcome!)
- [x] HF convert script update
- [ ] More robust conversions + better meta data