@Dampfinchen 2023-12-11
Awesome work!

In case you haven't seen it, Mistral also released an official inference code for vLLM https://github.com/vllm-project/vllm/pull/2011 including optimized CUDA kernels. I wonder if this could be useful for llama.cpp as well.

25 GB is a lot compared to 13B models, but since it's just 2 experts are running at the same time, I guess it would be as fast as 13B? Really wonder how it would run on my 32 GB+ 6 GB VRAM laptop. 13B is quite fast thanks to GPU offloading.