@slaren 2023-12-11
Generation speed should be similar to 13B, however prompt processing speed currently is very bad. Essentially, since every token can use different experts, each token in the batch is processed with a different matrix multiplication, and as a result the performance during prompt processing is not much better than during generation. We can improve this by grouping the tokens by the experts that they use and evaluating them together, but this is not yet implemented.