@EwoutH 2023-12-11
> Essentially, since every token can use different experts

Does that mean different experts need to be loaded in VRAM every token (if the VRAM isn’t large enough to hold all the experts simultaneously)?