@slaren 2023-12-11
> Does that mean different experts need to be loaded in VRAM every token (if the VRAM isn’t large enough to hold all the experts simultaneously)?

Each layer has its own set of experts. All the experts of the layers offloaded need to be loaded in VRAM. If the model doesn't fit in VRAM, you can reduce the number of layers offloaded with `-ngl` as usual.