@bergkvist 2023-12-11
q8_0 quantization worked quite well for me:
```
./quantize ./models/mixtral-8x7b-32k/ggml-model-f16.gguf \
           ./models/mixtral-8x7b-32k/ggml-model-q8_0.gguf \
           q8_0
```

I'm on an M2 Max macbook with 96GB of RAM, and this seemed to use around ~50GB.