@TheBloke 2023-12-11
Amazing work slaren!

I'm uploading GGUFs here: https://huggingface.co/TheBloke/Mixtral-8x7B-v0.1-GGUF