@TheBloke 2023-12-11
FYI if you didn't see it already yet, the official Mixtral is here now at : https://huggingface.co/mistralai/Mixtral-8x7B-v0.1/tree/main

They include the consolidated.pth files, but they're split in to multiple files because of the HF file limit.  And this PR doesn't currently support that, as it expects a single file and complains it doesn't know which of the .pth files to use

So I made my GGUFs from the torrent version like you recommended