@TheBloke 2023-12-11
Do I understand correctly that you do not quant the gate layers? They're F16?

That's good to know, because GPTQ is currently broken and one theory for how to fix it is not to quant the gate layers. So good to know that matches with how you went about it