@slaren 2023-12-11
> Do I understand correctly that you do not quant the gate layers? They're F16?

Yes, that's correct. @ggerganov implemented that, and I am not sure if there were any specific issues when quantizing these layers, but they are very small anyway and it barely increases the model size keeping them as f16.