@ggerganov 2023-12-11
Quantizing the gates works, but from very brief PPL tests, it showed slight improvement to not quantize them and the disk space increase is pretty small, so I decided to go with F16.

There is a lot to explore in regards to quantizing this arch - have no idea what is the best approach yet.