@TheBloke 2023-12-11
Great, thanks for the details.

I mistakenly made all k-quants at first and then realised that eg Q4_K_M and Q4_K_S were the same size, same with all the sub-variants

So now I've removed the size dupes to have:
* Q2_K
* Q3_K
* Q4_K
* Q5_K
* Q6_K

Plus Q4_0, Q5_0 and Q8_0.