@sroecker 2023-12-11
Interesting, trying to run this PR with a self-quantized Q3_K_M version of the torrent release on 2x12GB GPUs fails at: GGML_ASSERT: llama.cpp:3078: hparams.n_expert > 0
TheBloke's version works on CPU only, but otherwise fails at
GGML_ASSERT: ggml-cuda.cu:7680: false
which looks like multiple GPUs are not supported.