@arch-btw 2023-12-11
It works on 100% CPU for me.

Though I think I'm missing the correct prompt format, as it's just finishing my sentence.

> ./main -m models/mixtral-8x7b-v0.1.Q4_0.gguf -t 8 --color -c 32768 --temp 0.7 --repeat_penalty 1.1 -n -1 -p "Tell me what a car is"

Output:

>  **Tell me what a car is** made of, and I’ll tell you if it can run.

Anyway, very impressive. It runs about as fast as a 13B model for me.