@iibw 2023-12-11
HuggingFace has its own implementation now https://github.com/huggingface/transformers/commit/accccdd0087263a1e494e9c9ec30a43043ff3905