@DenisSergeevitch 2023-12-11
I don't know it is a Mac Metal issue or the Q5, but following instructions does not work properly (they are ignored). Has anyone encountered this?