@slaren 2023-12-11
@TetrisBlack @sroecker The issue with multi GPU should be fixed now, thanks for testing.