@happycube 2023-12-11
I think the next thing to do (not that I know how!) would be to allow offload of the 2B shared weights to the GPU... that would fit 25% of the active weights on 8GB GPU's in any (or no!) quantization and could be a nice speedup.