@ggerganov 2023-12-12
I think there might be a subtle bug in the Metal calculation.

Testing with the following command:

```bash
make -j && ./main -m models/mixtral-instruct-8x7b-fast/ggml-model-f16.gguf -ngl 1 --color --temp 0 --repeat_penalty 1 -c 4096 -n -1 -p "[INST] Prove that sqrt(2) is a rational number. [/INST]" --verbose-prompt -t 1
```

This is using the F16 model and the generation terminates too early:

```
system_info: n_threads = 1 / 24 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 | 

main: prompt: '[INST] Prove that sqrt(2) is a rational number. [/INST]'
main: number of tokens in prompt = 21
     1 -> ''
   733 -> ' ['
 16289 -> 'INST'
 28793 -> ']'
  1133 -> ' Pro'
   333 -> 've'
   369 -> ' that'
 18328 -> ' sq'
  2576 -> 'rt'
 28732 -> '('
 28750 -> '2'
 28731 -> ')'
   349 -> ' is'
   264 -> ' a'
 15202 -> ' rational'
  1474 -> ' number'
 28723 -> '.'
   733 -> ' ['
 28748 -> '/'
 16289 -> 'INST'
 28793 -> ']'

sampling: 
	repeat_last_n = 64, repeat_penalty = 1.000, frequency_penalty = 0.000, presence_penalty = 0.000
	top_k = 40, tfs_z = 1.000, top_p = 0.950, min_p = 0.050, typical_p = 1.000, temp = 0.000
	mirostat = 0, mirostat_lr = 0.100, mirostat_ent = 5.000
sampling order: 
CFG -> Penalties -> top_k -> tfs_z -> typical_p -> top_p -> min_p -> temp 
generate: n_ctx = 4096, n_batch = 512, n_predict = -1, n_keep = 0


 [INST] Prove that sqrt(2) is a rational number. [/INST] I cannot prove that sqrt(2) is a rational number, because it is not true. A rational number can be written as the quotient of two integers, while sqrt(2) is an irrational number. Its value cannot be expressed as a fraction.

Here's a simple proof to show that sqrt(2) is an irrational number:

Assume that sqrt(2) is a rational number. Then it can be written in the form p/q, where p and q are integers and q is not equal to 0, and the fraction is in its simplest form, i.e., p and q have no common factors other than 1.
 [end of text]

llama_print_timings:        load time =    4254.47 ms
llama_print_timings:      sample time =       3.44 ms /   149 runs   (    0.02 ms per token, 43313.95 tokens per second)
llama_print_timings: prompt eval time =     585.17 ms /    21 tokens (   27.87 ms per token,    35.89 tokens per second)
llama_print_timings:        eval time =    6648.71 ms /   148 runs   (   44.92 ms per token,    22.26 tokens per second)
llama_print_timings:       total time =    7255.08 ms
```

This does not happen when running CPU-only (`-ngl 0`):

<details>

```
system_info: n_threads = 1 / 24 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 | 

main: prompt: '[INST] Prove that sqrt(2) is a rational number. [/INST]'
main: number of tokens in prompt = 21
     1 -> ''
   733 -> ' ['
 16289 -> 'INST'
 28793 -> ']'
  1133 -> ' Pro'
   333 -> 've'
   369 -> ' that'
 18328 -> ' sq'
  2576 -> 'rt'
 28732 -> '('
 28750 -> '2'
 28731 -> ')'
   349 -> ' is'
   264 -> ' a'
 15202 -> ' rational'
  1474 -> ' number'
 28723 -> '.'
   733 -> ' ['
 28748 -> '/'
 16289 -> 'INST'
 28793 -> ']'

sampling: 
	repeat_last_n = 64, repeat_penalty = 1.000, frequency_penalty = 0.000, presence_penalty = 0.000
	top_k = 40, tfs_z = 1.000, top_p = 0.950, min_p = 0.050, typical_p = 1.000, temp = 0.000
	mirostat = 0, mirostat_lr = 0.100, mirostat_ent = 5.000
sampling order: 
CFG -> Penalties -> top_k -> tfs_z -> typical_p -> top_p -> min_p -> temp 
generate: n_ctx = 4096, n_batch = 512, n_predict = -1, n_keep = 0


 [INST] Prove that sqrt(2) is a rational number. [/INST] I apologize for the confusion, but sqrt(2) is actually not a rational number. A rational number can be expressed as the quotient of two integers, where the denominator is not equal to zero. However, sqrt(2) is an irrational number, which means it cannot be expressed as a simple fraction.

To prove that sqrt(2) is irrational, we can use proof by contradiction. Let's assume that sqrt(2) is rational, so it can be written as an irreducible fraction a/b, where a and b are coprime integers (no common factors other than 1), and b ≠ 0.

sqrt(2) = a/b

Squaring both sides, we get:

2 = a^2/b^2

Rearranging, we get:

a^2 = 2 * b^2

This equation tells us that a^2 is an even number because it equals 2 times another integer (b^2). Since a^2 is even, then a must also be even (only even numbers squared give even numbers). So, we can write a as 2 * k, where k is an integer.

Substituting this into our equation, we get:

(2 * k)^2 = 2 * b^2
4 * k^2 = 2 * b^2
b^2 = 2 * k^2

Now, we see that b^2 is also an even number, which implies that b is also even. But this contradicts our initial assumption that a and b are coprime, as they now share a common factor of 2. Therefore, our initial assumption that sqrt(2) is rational must be false. Thus, sqrt(2) is an irrational number. [end of text]

llama_print_timings:        load time =    4892.50 ms
llama_print_timings:      sample time =       9.65 ms /   402 runs   (    0.02 ms per token, 41653.71 tokens per second)
llama_print_timings: prompt eval time =   20577.94 ms /    21 tokens (  979.90 ms per token,     1.02 tokens per second)
llama_print_timings:        eval time =  197123.79 ms /   401 runs   (  491.58 ms per token,     2.03 tokens per second)
llama_print_timings:       total time =  217767.56 ms
```
</details>

Also using the quantum models with Metal backend leads to some subtle logical mistakes in the proof and sometimes typos, but it derives the proof correctly with CPU-only. These are anecdotal examples, but I'm worried that there might be an actual issue somewhere.

Does anyone else have similar observations?